Revert "Feat/socket.io mongodb adapter"
