+++ mongodb/models/user.ts
@@ -12,31 +12,11 @@ const UserSchema = new mongoose.Schema(
       required: true,
       trim: true,
       unique: true,
-      validate(value: string) {
-        if (
-          !/^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(
-            value
-          )
-        ) {
-          throw new Error('Invalid email');
-        }
-      },
     },
     password: {
       type: String,
       trim: true,
       minlength: 8,
-      validate(value: string) {
-        if (!value.match(/\d/) || !value.match(/[a-zA-Z]/)) {
-          throw new Error(
-            'Password must contain at least one letter and one number'
-          );
-        }
-      },
-    },
-    online: {
-      type: Boolean,
-      default: false,
     },
     lastActive: {
       type: Date,
