+++ package.json
@@ -2,6 +2,7 @@
   "name": "messenger-backend",
   "version": "1.0.0",
   "description": "",
+  "main": "index.js",
   "scripts": {
     "start": "node index.js",
     "build": "rimraf ./dist && tsc",
@@ -16,7 +17,6 @@
     "typescript": "^5.4.5"
   },
   "dependencies": {
-    "@socket.io/mongo-emitter": "^0.2.0",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
